Refactoring: extract functions responsible for running formulas.
This is a prefactoring for the coming fixes to the mechanism of running
formulas, see discussion in #767.

* Make JellyForm component into a directory
* Split out `runFormulas` and friends
* Add tests to describe current behavior

Change-type: patch
